Changed config and data readout
changed config to have title values and each sensor is referred to by a number. This makes it easier for the ESP32 to send an integer character for each sensor rather than a whole word.

Changed server to be outputting fake data for now while ESP32 takes a break.

Implemented hacky canvas data reduction which skips data points if window selects number of elements above a certain threshhold. Works well for a bit until lots of data is sent which happens quickly.